refactor: move get_stats() to Cachify_Backend interface
All backends implement this method and it is called unconditionally, so
let's move it to the interface.
The returned value is always parsed to integer, so we can unify the
return type, too.

+++ inc/class-cachify-hdd.php
@@ -142,12 +142,12 @@ final class Cachify_HDD implements Cachify_Backend {
 	/**
 	 * Get the cache size
 	 *
-	 * @return int Directory size
+	 * @return integer Cache size in bytes.
 	 *
 	 * @since 2.0
 	 */
 	public static function get_stats(): int {
-		return self::_dir_size( CACHIFY_CACHE_DIR );
+		return (int) self::_dir_size( CACHIFY_CACHE_DIR );
 	}
 
 	/**
